+++ UltimateCastbars/Backend/Modules/Debug.lua
@@ -55,19 +55,19 @@ end
 
 
 function Debug:StartDebug()
-  local playerName = UCB.charName
+  local playerName = UCB.unitNames.player
   local addonList = self:GetEnabledAddons(false, playerName) -- get all enabled addons except itself
   local cfg = UCB.GetValueConfig("debug")
   cfg.enabled = true
   cfg._addonList = addonList
-  DisableAddonList(addonList, UCB.charName)
+  DisableAddonList(addonList, UCB.unitNames.player)
 end
 
 function Debug:StopDebug()
     local cfg = UCB.GetValueConfig("debug")
     local addonList = cfg._addonList or {}
     if #addonList > 0 and cfg.enabled then
-        EnableAddonList(addonList, UCB.charName)
+        EnableAddonList(addonList, UCB.unitNames.player)
         cfg.enabled = false
         cfg._addonList = {}
     else